+++ CHANGELOG.md
@@ -1,26 +1,50 @@
+## 0.0.8
+
+- Features
+
+- Add sa-green color 
+- Enhance CustomModal with icon and improved error handling
+
+## 0.0.7
+
+- Features
+
+- Refactored Custom Search Dropdown Field to have an scrollable list
+
+## 0.0.6
+
+- Features
+
+* Added timer to pinput field
+* Refactor fields to match Figma design
+
 ## 0.0.5
 
-* Features
-- Added colors and images
+- Features
+
+* Added colors and images
 
 ## 0.0.4
 
-* Features
-- Added phone validator input using 'pinput' package
-- Refactor various fields to use 'reactive_forms' package
+- Features
+
+* Added phone validator input using 'pinput' package
+* Refactor various fields to use 'reactive_forms' package
 
 ## 0.0.3
 
-* Features
-- Added CustomSearchFormfield
+- Features
+
+* Added CustomSearchFormfield
 
 ## 0.0.2
 
-* Features
-- Added custom form fields with auto validation
-- Added custom button
-- Added custom header
+- Features
+
+* Added custom form fields with auto validation
+* Added custom button
+* Added custom header
 
 ## 0.0.1
 
-* Initial release
+- Initial release